data-generators/requirements.txt
kafka-python==2.0.2
numpy==1.20.1
